moved compute_samplex_arguments and make_samplex_arguments
