+++ qiskit_ibm_runtime/executor_estimator/prepare.py
@@ -15,31 +15,24 @@
 from __future__ import annotations
 
 import logging
-from collections import defaultdict
-from typing import TYPE_CHECKING, Any
+from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from collections.abc import Sequence
 
-    import numpy.typing as npt
-    from qiskit import QuantumCircuit
     from qiskit.primitives.containers.estimator_pub import EstimatorPub
-    from samplomatic.samplex import Samplex
 
     from ..options_models.measure_noise_learning_options import MeasureNoiseLearningOptions
     from ..options_models.twirling_options import TwirlingOptions
 
-import numpy as np
-from qiskit.quantum_info import Pauli, PauliList
-from samplomatic import ChangeBasis, build
-from samplomatic.utils import get_annotation
+from samplomatic import build
 
 from ..exceptions import IBMInputValueError
 from ..executor.calculate_twirling_shots import calculate_twirling_shots
 from ..quantum_program import QuantumProgram
 from ..quantum_program.quantum_program import SamplexItem
 from .trex_utils import create_trex_calibration_circuit
-from .utils import box_circuit, get_pauli_basis, pauli_to_ints, unbroadcast_index
+from .utils import box_circuit, compute_samplex_arguments, make_samplex_arguments
 
 logger = logging.getLogger(__name__)
 
@@ -154,144 +147,3 @@ def prepare(
     quantum_program._semantic_role = "estimator_v2"
 
     return quantum_program
-
-
-def compute_samplex_arguments(
-    pub: EstimatorPub,
-) -> tuple[npt.NDArray[float], npt.NDArray[int], list[tuple[tuple[int, ...], str]]]:
-    """Compute parameter values and basis changes to be used as inputs by the samplex.
-
-    To minimize the total number of circuits executions, this function takes the following
-    steps:
-        1. It creates a map between subsets of parameters and the observables that need to
-           be measured for each subset, applying broadcasting rules to params and observables.
-        2. It replaces the observables in that map with the minimal set of Pauli basis that
-           can be used to measure all such observables.
-        3. It flattens the map into two 1D arrays of equal lenght, containing the subsets of
-           parameters and basis changing gates respectively. When a subset of parameter maps
-           to more than one basis changing gate, the flattened array contains multiple copies
-           of it.
-
-    Overall, the two 1D arrays returned contain ``N`` elements, where ``N`` is the total number
-    of basis changing gates that need to be measured across all the different parameter sets.
-    Zipping them yields parameter–basis pairs, where each parameter value must be measured using
-    its associated change basis.
-
-    The two arrays have the format required by samplomatic and can be pass straight to the samplex
-    via ``samplex.inputs()``.
-
-    Args:
-        pub: An estimator PUB.
-
-    Return:
-        A tuple ``(flat_parameter_values, change_basis, param_basis_pairs)`` where:
-
-            * ``flat_parameter_values`` is a 1-D array of parameter values in the format expected
-            by ``samplex.inputs()``. The array is of length ``N``, the total number of
-            basis-changing gates required across all parameter sets.
-
-            * ``change_basis`` is a 1-D array of length ``N`` containing the basis-changing gates
-            associated with ``flat_parameter_values``, also in the format expected by
-            ``samplex.inputs()``.
-
-            * ``param_basis_pairs`` is a list of ``N`` tuples ``(ndindex, basis)`` describing the
-            correspondence between the two arrays. For the i-th tuple:
-                - ``ndindex`` is the N-dimensional index of the parameter entry in
-                    ``pub.parameter_values``.
-                - ``basis`` is the measurement basis associated with that parameter entry.
-    """
-    parameter_values = pub.parameter_values
-    observables = pub.observables
-    bcast_shape = pub.shape
-
-    # Step 1.
-    # Generate a map between param ndindices to pauli basis and the observable terms that they
-    # measure
-    param_obs_map: dict[set] = defaultdict(lambda: defaultdict(set))  # type: ignore[type-arg]
-    for bcast_index in np.ndindex(bcast_shape):
-        param_index = unbroadcast_index(bcast_index, parameter_values.shape)
-        obs = observables[unbroadcast_index(bcast_index, observables.shape)]
-        for obs_term, _ in obs.items():
-            pauli_basis = get_pauli_basis(obs_term)
-            param_obs_map[param_index][pauli_basis].add(obs_term)
-
-    # Step 2.
-    # Collect the Paulis to measure for each parameter value in commuting sets
-    param_meas_groups_map = {}
-    for param_index, pauli_map in param_obs_map.items():
-        pauli_set = list(pauli_map)
-        meas_groups = PauliList(pauli_set).group_commuting(qubit_wise=True)
-        param_meas_groups_map[param_index] = meas_groups
-
-    # Figure out measurement Pauli basis for each set of commuting Paulis
-    param_basis_map = {}
-    for param_index, meas_groups in param_meas_groups_map.items():
-        param_basis_map[param_index] = [
-            Pauli((np.logical_or.reduce(paulis.z), np.logical_or.reduce(paulis.x)))
-            for paulis in meas_groups
-        ]
-
-    # Step 3. Flatten the params.
-    # We flatten params into a 1D array and generate a corresponding 1D `change_basis` array. Both
-    # arrays contain ``num_basis`` elements.
-    num_basis = sum(len(basis) for basis in param_basis_map.values())
-    flat_parameter_values = np.empty(
-        (num_basis, parameter_values.num_parameters),
-        dtype=float,
-    )
-    change_basis = np.empty((num_basis, observables.num_qubits), dtype=int)
-
-    basis_idx = 0
-    for ndindex, basis in param_basis_map.items():
-        for bases in basis:
-            change_basis[basis_idx] = pauli_to_ints(bases)
-            flat_parameter_values[basis_idx] = parameter_values.as_array()[ndindex]
-            basis_idx += 1
-
-    # Step 4. Log info.
-    param_basis_pairs: list[tuple[tuple[int, ...], str]] = [
-        (ndindex, bases.to_label()) for ndindex, basis in param_basis_map.items() for bases in basis
-    ]
-
-    return flat_parameter_values, change_basis, param_basis_pairs
-
-
-def make_samplex_arguments(
-    samplex: Samplex,
-    boxed_circuit: QuantumCircuit,
-    flat_parameter_values: npt.NDArray[float],
-    change_basis: npt.NDArray[int],
-) -> dict[str, Any]:
-    """Build a samplex args dictionary consisting of ``change_basis`` and parameters data.
-
-    Args:
-        samplex: A samplex object to create args to.
-        boxed_circuit: A boxed circuit related to the samplex.
-        flat_parameter_values: A flattened array of parameter values.
-        change_basis: An array of bases to change.
-
-    Returns:
-        A samplex args dictionary.
-    """
-    # Prepare samplex_arguments
-    samplex_arguments = {}
-    if samplex.inputs().get_specs("parameter_values"):
-        samplex_arguments["parameter_values"] = flat_parameter_values
-
-    # Set changing basis gates
-    for spec in samplex.inputs().get_specs("basis_changes"):
-        # Default to np.zeros, to ensure that every mid-circuit measurement
-        # that may be present is performed without basis changing gates
-        samplex_arguments[spec.name] = np.zeros(spec.shape)
-
-    # Finalize basis changing gates for the final measurements
-    for instr in boxed_circuit.reverse_ops():
-        op = instr.operation
-        if op.name == "box" and (change_basis_annot := get_annotation(op, ChangeBasis)):
-            samplex_arguments[f"basis_changes.{change_basis_annot.ref}"] = change_basis
-            break
-    else:
-        # This should not be reachable
-        raise ValueError("Could not find a change basis annotation.")
-
-    return samplex_arguments
