+++ qiskit_ibm_runtime/executor_estimator/zne/zne_utils.py
@@ -35,9 +35,8 @@ from ...executor.calculate_twirling_shots import calculate_twirling_shots
 from ...options_models.zne_options import ZNE_DEFAULT_NOISE_FACTORS
 from ...quantum_program import QuantumProgram
 from ...quantum_program.quantum_program import SamplexItem
-from ..prepare import compute_samplex_arguments, make_samplex_arguments
 from ..trex_utils import create_trex_calibration_circuit
-from ..utils import box_circuit
+from ..utils import box_circuit, compute_samplex_arguments, make_samplex_arguments
 from .gate_folding import GateFolding
 
 logger = logging.getLogger(__name__)
