+++ test/unit/executor_estimator/test_estimator_v2_prepare.py
@@ -23,7 +23,8 @@ from qiskit.primitives.containers.estimator_pub import EstimatorPub, Observables
 from qiskit.quantum_info import SparsePauliOp
 
 from qiskit_ibm_runtime.exceptions import IBMInputValueError
-from qiskit_ibm_runtime.executor_estimator.prepare import compute_samplex_arguments, prepare
+from qiskit_ibm_runtime.executor_estimator.prepare import prepare
+from qiskit_ibm_runtime.executor_estimator.utils import compute_samplex_arguments
 from qiskit_ibm_runtime.options_models.measure_noise_learning_options import (
     MeasureNoiseLearningOptions,
 )
